refactor(auth): move RbacRaw
RHINENG-1304
