+++ src/main/java/com/redhat/cloud/policies/app/auth/models/RbacRaw.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.redhat.cloud.policies.app.auth;
+package com.redhat.cloud.policies.app.auth.models;
 
 import java.util.List;
 import java.util.Map;
